+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ TBD
   * REMOVED: qspi_io based RTOS flash driver
   * ADDED: lib_qspi_fast_read and lib_quadflash based RTOS flash driver
   * ADDED: RTOS flash driver function to set core affinity for QSPI transactions
+  * ADDED: New API to specify raw or nibble swapped individual qspi flash reads
 
 2.0.0
 -----